Added some sanity checks in samples to avoid failing cases.

+++ samples/j2ee/com.ibm.sbt.sample.web/WebContent/samples/java/Social/Activities/Delete ActivityNode.jsp
@@ -42,15 +42,19 @@
 	<%
 	try {		
 		ActivityService activityService = new ActivityService();
-		Activity activity = activityService.getMyActivities().get(0);
+		ActivityList activities = activityService.getMyActivities();
 		
-		ActivityNode actNode = new ActivityNode(activityService, activity.getActivityId());
-		actNode.setEntryType(ActivityNodeType.Chat.getActivityNodeType());
-		actNode.setTitle("ActivityNode for Deletion" + System.currentTimeMillis());
-		actNode.setContent("ActivityNodeContent");
-		actNode = activityService.createActivityNode(actNode);
-		activityService.deleteActivityNode(actNode.getActivityId());
-		out.println("Activity Node Created and Deleted : " + actNode.getTitle());
+		if(activities != null && !activities.isEmpty()) {
+			ActivityNode actNode = new ActivityNode(activityService, activities.get(0).getActivityId());
+			actNode.setEntryType(ActivityNodeType.Chat.getActivityNodeType());
+			actNode.setTitle("ActivityNode for Deletion" + System.currentTimeMillis());
+			actNode.setContent("ActivityNodeContent");
+			actNode = activityService.createActivityNode(actNode);
+			activityService.deleteActivityNode(actNode.getActivityId());
+			out.println("Activity Node Created and Deleted : " + actNode.getTitle());
+		}  else {
+			out.println("No Activites Found");
+		}
 	} catch (Throwable e) {
 		out.println("<pre>");
 		out.println(e.getMessage());
