+++ samples/j2ee/com.ibm.sbt.sample.web/WebContent/samples/java/Social/Activities/Delete Member.jsp
@@ -14,6 +14,7 @@
  * permissions and limitations under the License.
  */-->
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
+<%@page import="com.ibm.commons.util.StringUtil"%>
 <%@page import="com.ibm.sbt.services.client.connections.activity.Member"%>
 <%@page import="com.ibm.sbt.services.client.connections.activity.MemberList"%>
 <%@page import="java.util.Date"%>
@@ -43,18 +44,26 @@
 	try {		
 		ActivityService activityService = new ActivityService();
 		ActivityList activities = activityService.getMyActivities();
-		Activity activity = activities.get(0);
-		MemberList members = activityService.getMembers(activity.getActivityId());
-		String memberToBeDeleted = "" ;
-		for(Member mem : members) {
-			if(mem.getRole() != "owner") {
-				memberToBeDeleted = mem.getMemberId();
-				break;
+		if(activities != null && !activities.isEmpty()) {
+			Activity activity = activities.get(0);
+			MemberList members = activityService.getMembers(activity.getActivityId());
+			String memberToBeDeleted = "" ;
+			for(Member mem : members) {
+				if(mem.getRole() != "owner") {
+					memberToBeDeleted = mem.getMemberId();
+					break;
+				}
 			}
+			
+			if(StringUtil.isEmpty(memberToBeDeleted)) { 
+				out.println("No Member to be Deleted.");
+			} else {
+				activityService.deleteMember(activity.getActivityId(), memberToBeDeleted);
+				out.println("Member deleted : " + memberToBeDeleted);
+			}
+		}  else {
+			out.println("No Activites Found");
 		}
-		
-		activityService.deleteMember(activity.getActivityId(), memberToBeDeleted);
-		out.println("Member deleted : " + memberToBeDeleted);
 	} catch (Throwable e) {
 		out.println("<pre>");
 		out.println(e.getMessage());
